state: rename portalTrieDB file
