+++ packages/portalnetwork/src/networks/state/portalTrieDB.ts
